V1.1.9.9 Fix a bug when index file large then 2GB and using RamIndex, it will raise a exception said can't set System.IO.FileStream.set_Position a negative number.

+++ C#/src/Hubble.Framework/Hubble.Framework/IO/CachedFileStream.cs
@@ -280,7 +280,7 @@ namespace Hubble.Framework.IO
             {
                 using (System.IO.FileStream fs = new FileStream(_FilePath, FileMode.Open, FileAccess.Read, FileShare.Read))
                 {
-                    fs.Position = index * CachedFileBufferManager.BufferUnitSize;
+                    fs.Position = (long)index * CachedFileBufferManager.BufferUnitSize;
                     int offset = 0;
                     int read = 0;
